% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_datatable.R
\name{aggregate_datatable}
\alias{aggregate_datatable}
\title{aggregates data.tables with two columns: 1st col Country/Industry whatever, 2nd col "V1" with values. 
First, the data.table is sorted based on the values of column "V1". Then, all but n rows are aggregated to one sector "Rest".}
\usage{
aggregate_datatable(dt, n, col)
}
\arguments{
\item{dt}{data.table with two columns: 1st col Country/Industry whatever, 2nd col "V1" with values}

\item{n}{number of sectors/countries/etc. that stay}
}
\value{
a aggregated dtrix
}
\description{
aggregates data.tables with two columns: 1st col Country/Industry whatever, 2nd col "V1" with values. 
First, the data.table is sorted based on the values of column "V1". Then, all but n rows are aggregated to one sector "Rest".
}

