% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_matrix.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{aggregates matrices.}
\usage{
aggregate_matrix(mat, n, fun = mean)
}
\arguments{
\item{mat}{matrix with years as rownames and sectors/countries(/...) as colnames --> output of function cumulate_matrix with order = T}

\item{n}{number of sectors/countries/etc. that stay}

\item{fun}{function,  determines how first n sectors/countries/etc are determined}
}
\value{
a aggregated matrix
}
\description{
aggregates matrices.
}

