% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_timeseries.R
\name{create_timeseries}
\alias{create_timeseries}
\title{creates a data.table which contains the timeseries of each coefficient of a timeseries of matrices}
\usage{
create_timeseries(mat.list)
}
\arguments{
\item{mat.list}{a list containing a timeseries of matrices. Each element of the list represents on matrix of year t.}
}
\description{
creates a data.table which contains the timeseries of each coefficient of a timeseries of matrices
}

