% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate_matrix.R
\name{cumulate_matrix}
\alias{cumulate_matrix}
\title{colwise cumulate values of a matrix (required for stacked_area_plot)}
\usage{
cumulate_matrix(mat, order = T)
}
\arguments{
\item{mat}{matrix}

\item{order}{boolean}
}
\value{
a cumulated matrix
}
\description{
colwise cumulate values of a matrix (required for stacked_area_plot)
}

