% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{makes a cross-validation of a time-series. The time-series is split at a given year into a training data set and a test data set. A model is fitted to the training data and forecasts are made to the remaining years. These forecasts are compared to the "real" values.}
\usage{
cv(x, years.obs, year.split, type = "glm")
}
\arguments{
\item{x}{a timeseries of length(years.obs) observations}

\item{years.obs}{years for which observations exist}

\item{years.split}{first year, where time-series should be splitted to training and test data set. year.split is the last year of the first training data set.}
}
\description{
makes a cross-validation of a time-series. The time-series is split at a given year into a training data set and a test data set. A model is fitted to the training data and forecasts are made to the remaining years. These forecasts are compared to the "real" values.
}

