% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_A_column.R
\name{extrapolate_A_column}
\alias{extrapolate_A_column}
\title{Returns a list containing the extrapolated matrices.}
\usage{
extrapolate_A_column(col.list, years.obs, years.est, scaleTo1 = F,
  parallel = F, n.cores = 1, save.output = F, dir)
}
\arguments{
\item{col.list}{a list containing a timeseries of matrices. Each element of the list represents on matrix of year t.}

\item{years.obs}{years for which observations exist. Needs to be equal to number of columns of ts.dt}

\item{years.est}{years for which estimations/extrapolations should be done}

\item{parallel}{boolean. calculation parallel on multiple cores}

\item{n.cores}{number of cores. Default n.cores = 1.}

\item{save.output}{boolean. Shall the output (new matrices) be saved as .csv to directory dir (specified with param dir)?}

\item{dir}{directory to where the new matrices should be saved}
}
\description{
Returns a list containing the extrapolated matrices.
}

