% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_matrix.R
\name{extrapolate_matrix}
\alias{extrapolate_matrix}
\title{Returns a list containing the extrapolated matrices.}
\usage{
extrapolate_matrix(mat.list, years.obs, years.est, parallel = F,
  n.cores = 1, save.output = F, dir)
}
\arguments{
\item{years.obs}{years for which observations exist. Needs to be equal to number of columns of ts.dt}

\item{years.est}{years for which estimations/extrapolations should be done}

\item{parallel}{boolean. calculation parallel on multiple cores}

\item{n.cores}{number of cores. Default n.cores = 1.}

\item{save.output}{boolean. Shall the output (new matrices) be saved as .csv to directory dir (specified with param dir)?}

\item{dir}{directory to where the new matrices should be saved}

\item{mat_list}{a list containing a timeseries of matrices. Each element of the list represents on matrix of year t.}
}
\description{
Returns a list containing the extrapolated matrices.
}

