% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{fitmodel}
\alias{fitmodel}
\title{fits a model to a timeseries and extrapolates it into the future}
\usage{
fitmodel(x, years.obs, years.est, type = "glm")
}
\arguments{
\item{x}{a timeseries of length(years.obs) observations}

\item{years.obs}{years for which observations exist}

\item{years.est}{years for which estimations/extrapolations should be done}
}
\description{
fits a model to a timeseries and extrapolates it into the future
}

