% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatrix4Country.R
\name{getMatrix4Country}
\alias{getMatrix4Country}
\title{function to get detailed matrix for ONE country.}
\usage{
getMatrix4Country(mat, country.name, time.vec)
}
\arguments{
\item{mat}{matrix}

\item{country.name}{character, name of the country}

\item{time.vec}{vector}
}
\value{
a matrix
}
\description{
function to get detailed matrix for ONE country.
}

