% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plots the time series of one cell/coefficient of a list that contains time series of matrices (such as SUT, Y, etc.).}
\usage{
plot_timeseries(mat.list, row.index, col.index, years.obs, years.est = F, ...)
}
\arguments{
\item{row.index}{the row index of the cell}

\item{col.index}{the col.index of the cell}

\item{...}{further arguments for plot(). See also ?plot}

\item{mat_list}{a list containing a timeseries of matrices/2D-df/dt. Each element of the list represents on matrix of year t.}

\item{years}{A vector with the years for which data exists. Needs to be equal to the length of @param mat_list}
}
\description{
plots the time series of one cell/coefficient of a list that contains time series of matrices (such as SUT, Y, etc.).
}

