% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleTo100.R
\name{scaleTo100}
\alias{scaleTo100}
\title{scales matrix to 100.}
\usage{
scaleTo100(mat, baseyear)
}
\arguments{
\item{mat}{matrix}

\item{baseyear}{numeric, year that should be 100}
}
\value{
a matrix
}
\description{
scales matrix to 100.
}

