% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_area_plot.R
\name{stacked_area_plot}
\alias{stacked_area_plot}
\title{stacked area plot.}
\usage{
stacked_area_plot(mat, colorvec = 1:100, xaxt.steps, yaxt.steps, legend = T,
  legend.pos = "topleft", ...)
}
\arguments{
\item{mat}{matrix with years as rownames and sectors/countries(/...) as colnames --> output of function cumulate_matrix with order = T}

\item{colorvec}{vector with colorcodes (length(colorvec) = ncol(mat))}

\item{xaxt.steps}{steps for drawing tick marks at x-axis}

\item{yaxt.steps}{steps for drawing tick marks at x-axis}

\item{legend}{boolean,  draw a legend??}

\item{legend.pos}{coordinates (c(x,y) or position (e.g."bottomleft") of the legend}
}
\value{
a plot
}
\description{
stacked area plot.
}

