% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_forecast.R
\name{validate_forecast}
\alias{validate_forecast}
\title{.}
\usage{
validate_forecast(mat.list, years.obs, year.split, parallel = F,
  n.cores = 1, type = "glm", scale = F)
}
\arguments{
\item{years.obs}{years for which observations exist. Needs to be equal to number of columns of ts.dt}

\item{parallel}{boolean. calculating parallelly on multiple cores?}

\item{n.cores}{number of cores. Default n.cores = 1.}

\item{scale}{should each timeseries be scaled??}

\item{mat_list}{a list containing a timeseries of matrices. Each element of the list represents on matrix of year t.}

\item{years.split}{first year, where time-series should be splitted to training and test data set. year.split is the last year of the first training data set.}
}
\description{
.
}

