% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2mat.R
\name{vec2mat}
\alias{vec2mat}
\title{?}
\usage{
vec2mat(x, n.row, n.col)
}
\arguments{
\item{x}{a data.table with timeseries of all coefficients of a matrix. (Rowwise, i.e. 1st element if (1,1), 2nd is (1,2), etc. )}

\item{n.row}{number of rows of the output matrix}

\item{n.col}{number of cols of the output matrix}
}
\description{
?
}

