% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_area_plot.R
\name{stacked_area_plot}
\alias{stacked_area_plot}
\title{stacked area plot.}
\usage{
stacked_area_plot(data, agg = F, agg.level, colorvec = 1:100, xaxt.steps,
  yaxt.steps, ylim.adjust = 0, legend = T, legend.pos = "topleft",
  legend.ncol = 1, legend.bty = "o", ...)
}
\arguments{
\item{data}{either matrix with Years as rownames and sectors/countries(/...) as colnames --> output of function cumulate_matrix with order = T}

\item{agg}{aggregated? calls aggregate_matrix function. Boolean}

\item{agg.level}{aggregation level. Required if agg = T.}

\item{colorvec}{vector with colorcodes (length(colorvec) = ncol(mat))}

\item{xaxt.steps}{steps for drawing tick marks at x-axis}

\item{yaxt.steps}{steps for drawing tick marks at x-axis}

\item{legend}{boolean,  draw a legend??}

\item{legend.pos}{coordinates (c(x,y) or position (e.g."bottomleft") of the legend}
}
\value{
a plot
}
\description{
stacked area plot.
}

